requirements.txt
playwright==1.55.0
requests==2.31.0
openai
python-dotenv

